+++ SBBMobileDesignSystemSwiftUI/Colors/SBBColors.swift
@@ -51,6 +51,7 @@ public struct SBBColor {
     public static var textWhite: Color { color(withString: "textWhite") }
     public static var background: Color { color(withString: "background") }
     public static var buttonBackground: Color { color(withString: "buttonBackground") }
+    public static var disabledButtonBackground: Color { color(withString: "disabledButtonBackground") }
     public static var viewBackground: Color { color(withString: "viewBackground") }
     public static var divider: Color { color(withString: "divider") }
     public static var border: Color { color(withString: "border") }